v0.05 Nav+Marquee changes
Changed the logos in the marquee to made up companies and the size of them.
The navigation tab was changed to be always at the top of the page with rounded edges and a logo in the top left corner. (Also for now added a separate folder for logos)
Changed the text-box to a paragraph with a different RGB style borderaround it.
Changed the Title.
Deleted bugged/not needed pieces of code.
CSS: Adjusted fixed numbers to % to make the parts scale better after adding more reactive code.
Added @media to nav and banner2  (needs name change after a more permanent image)